+++ src/classes/interactions.ts
@@ -1,10 +1,11 @@
 import config from 'config';
-import { GuildQueue, Track } from 'discord-player';
+import { GuildQueue, PlayerTimestamp, Track } from 'discord-player';
 import {
     ApplicationCommandOptionChoiceData,
     AutocompleteInteraction,
     ChatInputCommandInteraction,
     EmbedAuthorOptions,
+    EmbedFooterData,
     GuildMember,
     Interaction,
     Message,
@@ -113,6 +114,71 @@ abstract class BaseInteraction {
         return thumbnailUrl;
     }
 
+    protected getFooterDisplayPageInfo(interaction: ChatInputCommandInteraction, queue: GuildQueue): EmbedFooterData {
+        if (!queue.tracks.data.length) {
+            return { text: 'Page 1 of 1 (0 tracks)' };
+        }
+
+        const pageIndex: number = (interaction.options.getNumber('page') || 1) - 1;
+        const totalPages: number = Math.ceil(queue.tracks.data.length / 10) || 1;
+        return {
+            text: `Page ${pageIndex + 1} of ${totalPages} (${queue.tracks.data.length} tracks)`
+        };
+    }
+
+    protected getDisplayTrackRequestedBy = (track: Track): string => {
+        return track.requestedBy ? `<@${track.requestedBy.id}>` : 'Unavailable';
+    };
+
+    protected getDisplayQueueProgressBar(queue: GuildQueue): string {
+        const timestamp: PlayerTimestamp = queue.node.getTimestamp()!;
+        let progressBar: string = `**\`${timestamp.current.label}\`** ${queue.node.createProgressBar({
+            queue: false,
+            length: this.playerOptions.progressBar.length ?? 12,
+            timecodes: this.playerOptions.progressBar.timecodes ?? false,
+            indicator: this.playerOptions.progressBar.indicator ?? '🔘',
+            leftChar: this.playerOptions.progressBar.leftChar ?? '▬',
+            rightChar: this.playerOptions.progressBar.rightChar ?? '▬'
+        })} **\`${timestamp.total.label}\`**`;
+
+        if (Number(queue.currentTrack?.raw.duration) === 0 || queue.currentTrack?.duration === '0:00') {
+            progressBar = '_No duration available._';
+        }
+
+        if (queue.currentTrack?.raw.live) {
+            progressBar = `${this.embedOptions.icons.liveTrack} **\`LIVE\`** - Playing continuously from live source.`;
+        }
+
+        return progressBar;
+    }
+
+    protected getDisplayRepeatMode(repeatMode: number, state: string = 'info'): string {
+        let loopModeUserString: string;
+        let icon: string;
+
+        switch (repeatMode) {
+            case 1:
+                loopModeUserString = 'track';
+                icon = state === 'info' ? this.embedOptions.icons.loop : this.embedOptions.icons.looping;
+                break;
+            case 2:
+                loopModeUserString = 'queue';
+                icon = state === 'info' ? this.embedOptions.icons.loop : this.embedOptions.icons.looping;
+                break;
+            case 3:
+                loopModeUserString = 'autoplay';
+                icon = state === 'info' ? this.embedOptions.icons.autoplay : this.embedOptions.icons.autoplaying;
+                break;
+            default:
+                return '';
+        }
+
+        return (
+            `**${icon} Looping**\n` +
+            `Loop mode is set to **\`${loopModeUserString}\`**. You can change it with **\`/loop\`**.\n\n`
+        );
+    }
+
     abstract execute(
         params: BaseInteractionParams
     ): Promise<Message<boolean> | ApplicationCommandOptionChoiceData | void>;
@@ -123,9 +189,9 @@ abstract class BaseInteractionWithEmbedResponse extends BaseInteraction {
         super();
     }
 
-    protected async getEmbedUserAuthor(
+    protected getEmbedUserAuthor(
         interaction: MessageComponentInteraction | ChatInputCommandInteraction
-    ): Promise<EmbedAuthorOptions> {
+    ): EmbedAuthorOptions {
         let authorName: string = '';
         if (interaction.member instanceof GuildMember) {
             authorName = interaction.member.nickname || interaction.user.username;
@@ -139,12 +205,13 @@ abstract class BaseInteractionWithEmbedResponse extends BaseInteraction {
         };
     }
 
-    protected async getEmbedQueueAuthor(
+    protected getEmbedQueueAuthor(
         interaction: MessageComponentInteraction | ChatInputCommandInteraction,
         queue: GuildQueue
-    ): Promise<EmbedAuthorOptions> {
+    ): EmbedAuthorOptions {
+        const bitrate = queue.channel ? queue.channel.bitrate / 1000 : 0;
         return {
-            name: `Channel: ${queue.channel!.name} (${queue.channel!.bitrate / 1000}kbps)`,
+            name: `Channel: ${queue.channel!.name} (${bitrate}kbps)`,
             iconURL: interaction.guild!.iconURL() || this.embedOptions.info.fallbackIconUrl
         };
     }