+++ src/interactions/commands/player/pause.ts
@@ -27,13 +27,13 @@ class PauseCommand extends BaseSlashCommandInteraction {
 
         const currentTrack: Track = queue.currentTrack!;
 
-        this.togglePauseState(queue, logger);
+        this.togglePauseState(logger, queue);
 
         logger.debug('Responding with success embed.');
         return await interaction.editReply({
             embeds: [
                 new EmbedBuilder()
-                    .setAuthor(await this.getEmbedUserAuthor(interaction))
+                    .setAuthor(this.getEmbedUserAuthor(interaction))
                     .setDescription(
                         `**${this.embedOptions.icons.pauseResumed} ${
                             queue.node.isPaused() ? 'Paused Track' : 'Resumed track'
@@ -45,7 +45,7 @@ class PauseCommand extends BaseSlashCommandInteraction {
         });
     }
 
-    private togglePauseState(queue: GuildQueue, logger: Logger): void {
+    private togglePauseState(logger: Logger, queue: GuildQueue): void {
         queue.node.setPaused(!queue.node.isPaused());
         logger.debug(`Set paused state to ${queue.node.isPaused()}.`);
     }