+++ src/interactions/components/filters-disable-button.ts
@@ -1,5 +1,6 @@
 import { GuildQueue, useQueue } from 'discord-player';
 import { EmbedBuilder } from 'discord.js';
+import { Logger } from 'pino';
 import { BaseComponentInteraction } from '../../classes/interactions';
 import { BaseComponentParams, BaseComponentReturnType } from '../../types/interactionTypes';
 import { checkQueueExists } from '../../utils/validation/queueValidator';
@@ -22,24 +23,32 @@ class FiltersDisableButtonComponent extends BaseComponentInteraction {
             checkQueueExists
         ]);
 
-        // Reset filters before enabling provided filters
-        if (queue.filters.ffmpeg.filters.length > 0) {
-            queue.filters.ffmpeg.setFilters(false);
-            logger.debug('Reset queue filters.');
-        }
+        this.resetFilters(queue, logger);
 
         logger.debug('Responding with success embed.');
         return await interaction.editReply({
             embeds: [
                 new EmbedBuilder()
-                    .setAuthor(await this.getEmbedUserAuthor(interaction))
+                    .setAuthor(this.getEmbedUserAuthor(interaction))
                     .setDescription(
-                        `**${this.embedOptions.icons.success} Disabled filters**\nAll audio filters have been disabled.`
+                        `**${this.embedOptions.icons.success} Disabled filters**\n` +
+                            'All audio filters have been disabled.'
                     )
                     .setColor(this.embedOptions.colors.success)
             ],
             components: []
         });
+    }
+
+    private resetFilters(queue: GuildQueue, logger: Logger): void {
+        if (queue.filters.ffmpeg.filters.length > 0) {
+            queue.filters.ffmpeg.setFilters(false);
+            logger.debug('Reset queue filters.');
+        }
+
+        if (queue.filters.biquad) {
+            queue.filters.biquad.disable();
+        }
     }
 }